1. remove pair_hf/pair_oai. 2. add continual_generate in BaseLLM
