+++ .gitignore
@@ -13,3 +13,4 @@ examples/check_results.py
 results/
 .vscode
 cmd_jbb.sh
+/test.py
