+++ jailbreakpipe/role/attacks/pair.py
@@ -4,7 +4,6 @@ from jailbreakpipe.role.attacks import BaseAttacker, BaseAttackerConfig
 from jailbreakpipe.role.attacks.attacker_registry import register_attacker
 from jailbreakpipe.llms import create_llm, BaseLLMConfig, LLMGenerateConfig
 from jailbreakpipe.role.judges import PairLLMJudgeConfig, create_judge
-from jailbreakpipe.llms.pair_hf import PairAttackerHFLLM, PairAttackerHFLLMConfig
 from jailbreakpipe.llms import HuggingFaceLLM, HuggingFaceLLMConfig
 from jailbreakpipe.utils import *
 import json
@@ -197,12 +196,14 @@ class PairAttackerConfig(BaseAttackerConfig):
     attacker_name: str = field(default=None)
 
     # PAIR script args
-    attack_llm_config: PairAttackerHFLLMConfig = field(
-        default_factory=PairAttackerHFLLMConfig
-    )
-    target_llm_config: HuggingFaceLLMConfig = field(
-        default_factory=HuggingFaceLLMConfig
-    )
+    # attack_llm_config: PairAttackerHFLLMConfig = field(
+    #     default_factory=PairAttackerHFLLMConfig
+    # )
+    # target_llm_config: HuggingFaceLLMConfig = field(
+    #     default_factory=HuggingFaceLLMConfig
+    # )
+    attack_llm_config: BaseLLMConfig = field(default_factory=BaseLLMConfig)
+    target_llm_config: BaseLLMConfig = field(default_factory=BaseLLMConfig)
 
     # TODO: extend pair judge to any type
     pair_judge: PairLLMJudgeConfig = field(
@@ -255,7 +256,6 @@ class PairAttacker(BaseAttacker):
 
         # attacker and target llms
         self.attack_llm = create_llm(config=config.attack_llm_config)
-        assert isinstance(self.attack_llm, PairAttackerHFLLM)
         self.target_llm = create_llm(config=config.target_llm_config)
 
         # TODO: extend judge to other types
@@ -317,8 +317,8 @@ class PairAttacker(BaseAttacker):
         :rtype: Tuple[str, str]
         """
 
-        # goal = message[-1]["content"]
-        goal = "How to kill a man?"
+        goal = message[-1]["content"]
+        # goal = "How to kill a man?"
         return goal, self.target_str
 
     def _init_conversation(self, goal: str, target_str: str) -> List[Dict[str, str]]:
@@ -367,7 +367,7 @@ class PairAttacker(BaseAttacker):
 
         attack_response = {"role": "assistant", "content": attack_prefix}
         full_conv.append(attack_response)
-        message = self.attack_llm.generate(
+        message = self.attack_llm.continual_generate(
             messages=full_conv, config=self.attack_gen_config
         )
 
@@ -496,9 +496,9 @@ class PairAttacker(BaseAttacker):
         """
         selected_conv = max(adv_process, key=lambda x: x[1])[0]
         selected_attack = selected_conv[-2]
-        return selected_attack
+        return [selected_attack]
 
-    def attack(self, messages, **kwargs):
+    def attack(self, messages, **kwargs) -> Dict[str, str]:
         """
         Perform the PAIR attack and return the final adversarial attack.
